US Mine Site Incident Data Analysis
A data analysis was done on the incident data at mine sites within the United States. The analysis visualises various types of mines that are more prone to incidents and the types of incidents that frequently occur. The analysis proposes a model to predict the chance of fatality incase of an incident based on the type of incident, the industry involved and the geographical location. The location data successfully used in the model points to the flaws in labour laws and safety standards within the states making them more prone to incurring fatalities incase of incidents.